requirements.txt
asgiref==3.8.1
astroid==2.3.3
colorama==0.4.3
Django==5.1.3
django-crispy-forms==2.3
django-filter==24.3
django-login-required-middleware==0.4
django-widget-tweaks==1.4.5
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.5.1
wrapt==1.11.2
beautifulsoup4==4.12.3
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
soupsieve==2.6
tzdata==2024.2

